requirements.txt
docker>=5.0.0
kubernetes>=27.2.0 # introduction of suspend in jobspec
flask>=2.0.0
natsort>=8.0.0
Flask-BasicAuth>=0.2.0
MarkupSafe>=2.1.5
apache-libcloud>=3.8.0